requirements.txt
langchain==1.2.0
langchain-core==1.2.3
langchain-openai==1.1.6
openai==2.13.0
python-dotenv==1.2.1
pydantic==2.12.5
playwright==1.48.0